requirements.txt
# -------------------------------
# Core Automation
# -------------------------------
selenium==4.20.0
Appium-Python-Client==3.1.0

# -------------------------------
# Test Framework
# -------------------------------
pytest==8.2.0
pytest-xdist==3.6.1         # parallel execution
pytest-html==4.1.1          # basic reporting
pytest-rerunfailures==14.0  # flaky test handling

# -------------------------------
# HTTP / API / Health Checks
# -------------------------------
requests==2.31.0

# -------------------------------
# Config Management
# -------------------------------
PyYAML==6.0.1
python-dotenv==1.0.1

# -------------------------------
# Logging & Utilities
# -------------------------------
loguru==0.7.2
tenacity==8.2.3             # retry logic (better than manual retry)

# -------------------------------
# Reporting (Advanced - Optional but Recommended)
# -------------------------------
allure-pytest==2.13.5

# -------------------------------
# Data Handling (Optional)
# -------------------------------
pandas==2.2.2

# -------------------------------
#  (Optional)
# -------------------------------
browserstack-local==1.2.0
jsonschema==4.22.0
selenium-wire==5.1.0   # network inspection

# -------------------------------
# Dev / Linting (Optional but useful)
# -------------------------------
flake8==7.0.0
black==24.4.2